caching full top profile results. some refactors included
