+++ presenters/http/http.go
@@ -44,6 +44,10 @@ func toProfileJson(profile *entities.Profile) *profileJson {
 func toNonFungibleTokensJson(nfts *[]entities.NonFungibleToken) *[]nonFungibleTokenJson {
 	nftsJson := []nonFungibleTokenJson{}
 
+	if nfts == nil {
+		return &nftsJson
+	}
+
 	for _, nft := range *nfts {
 		nftsJson = append(nftsJson, *toNonFungibleJson(&nft))
 	}
@@ -53,6 +57,10 @@ func toNonFungibleTokensJson(nfts *[]entities.NonFungibleToken) *[]nonFungibleTo
 
 func toFungibleTokensJson(tokens *[]entities.FungibleToken) *[]fungibleTokenJson {
 	tokensJson := []fungibleTokenJson{}
+
+	if tokens == nil {
+		return &tokensJson
+	}
 
 	for _, token := range *tokens {
 		tokensJson = append(tokensJson, *toFungibleJson(&token))
@@ -64,6 +72,10 @@ func toFungibleTokensJson(tokens *[]entities.FungibleToken) *[]fungibleTokenJson
 func toStatisticsJson(stats *[]entities.Statistic) *[]statisticJson {
 	statsJson := []statisticJson{}
 
+	if stats == nil {
+		return &statsJson
+	}
+
 	for _, stat := range *stats {
 		statsJson = append(statsJson, *toStatisticJson(&stat))
 	}
@@ -73,6 +85,10 @@ func toStatisticsJson(stats *[]entities.Statistic) *[]statisticJson {
 
 func toInteractionsJson(interactions *[]entities.Interaction) *[]interactionJson {
 	interactionsJson := []interactionJson{}
+
+	if interactions == nil {
+		return &interactionsJson
+	}
 
 	for _, interaction := range *interactions {
 		interactionsJson = append(interactionsJson, *toInteractionJson(&interaction))
