+++ usecases/get_top_profiles.go
@@ -2,6 +2,7 @@ package usecases
 
 import (
 	"context"
+	"sync"
 
 	"github.com/etheralley/etheralley-core-api/common"
 	"github.com/etheralley/etheralley-core-api/entities"
@@ -13,12 +14,46 @@ type GetTopProfilesInput struct {
 
 type IGetTopProfilesUseCase func(ctx context.Context, input *GetTopProfilesInput) *[]entities.Profile
 
-func NewGetTopProfilesUseCase(logger common.ILogger, cacheGateway gateways.ICacheGateway) IGetTopProfilesUseCase {
+func NewGetTopProfilesUseCase(
+	logger common.ILogger,
+	cacheGateway gateways.ICacheGateway,
+	resolveENSName IResolveENSNameUseCase,
+) IGetTopProfilesUseCase {
 	return func(ctx context.Context, input *GetTopProfilesInput) *[]entities.Profile {
-		logger.Info(ctx, "get top profiles usecase")
+		if profiles, err := cacheGateway.GetTopViewedProfiles(ctx); err == nil {
+			return profiles
+		}
 
-		cacheGateway.GetTopAddresses(ctx)
+		addresses, err := cacheGateway.GetTopViewedAddresses(ctx)
 
-		return &[]entities.Profile{}
+		if err != nil {
+			return &[]entities.Profile{}
+		}
+
+		var wg sync.WaitGroup
+		profiles := make([]entities.Profile, len(*addresses))
+
+		for i, address := range *addresses {
+			wg.Add(1)
+
+			go func(i int, address string) {
+				defer wg.Done()
+
+				ensName, _ := resolveENSName(ctx, &ResolveENSNameInput{
+					Address: address,
+				})
+
+				profiles[i] = entities.Profile{
+					Address: address,
+					ENSName: ensName,
+				}
+			}(i, address)
+		}
+
+		wg.Wait()
+
+		cacheGateway.SaveTopViewedProfiles(ctx, &profiles)
+
+		return &profiles
 	}
 }
